População brasileira, 2012

População por sexo e faixa etária. Municípios brasileiros, 2012.

Usage

POPBR12

Format

Banco de dados com estimativas populacionais por sexo e faixa etária para os municípios brasileiros em 2012. É o último arquivo com estimativas populacionais a esse nível de detalhamento publicado no repositório FTP do DATASUS. O banco é apresentado na forma em que foi baixado de ftp.datasus.gov.br/dissemin/publicos/IBGE/POP/POPBR12.zip.